backend/requirements.txt
anyio==3.5.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
dadata==21.10.1
Django==3.2
django-debug-toolbar==3.2.4
h11==0.12.0
httpcore==0.14.4
httpx==0.21.3
idna==3.3
Pillow==9.2.0
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
urllib3==1.26.8
environ>=1.0
psycopg2-binary==2.9.3
django-crispy-forms==1.14.0
Jinja2~=3.1.2
python-decouple==3.6
jsonfield==3.1.0
gunicorn==20.1.0
# redis-cli==5.0.7
django-redis
redis~=4.3.4
django-jsonfield~=1.4.1
django-environ==0.9.0
sentry_sdk==1.6.0
gevent==21.12.0
greenlet==1.1.2
django-import-export==2.8.0
djangorestframework==3.14.0